Refactor code to use symbols and packages
* dump.lisp: Remove quickload statements for 'alexandria', 'cl-dbi',
and 'lmdb'.  Replace 'use-package' with 'defpackage' and 'in-package'.
(plists->csv): Replace 'alexandria:plist-alist' with 'plist-alist'.
(fetch-results-from-sql): Replace 'dbi:with-connection',
'dbi:prepare', 'dbi:execute', 'dbi:fetch-all' with 'with-connection',
'prepare', and 'execute' respectively.
(store-sample-data-in-lmdb): Replace 'lmdb:with-env', 'lmdb:*env*',
'lmdb:get-db', 'lmdb:with-txn', 'lmdb:put' with 'with-env', '*env*',
'get-db', 'with-txn' and 'put' respectively.